feat: rename session components to section components in MyAccount pages
