+++ packages/core/src/pages/account/orders/[id].tsx
@@ -10,8 +10,8 @@ import CUSTOM_COMPONENTS from 'src/customizations/src/components'
 import type { MyAccountProps } from 'src/experimental/myAccountSeverSideProps'
 
 import { getGlobalSectionsData } from 'src/components/cms/GlobalSections'
-import { default as AfterSession } from 'src/customizations/src/myAccount/extensions/orders/[id]/after'
-import { default as BeforeSession } from 'src/customizations/src/myAccount/extensions/orders/[id]/before'
+import { default as AfterSection } from 'src/customizations/src/myAccount/extensions/orders/[id]/after'
+import { default as BeforeSection } from 'src/customizations/src/myAccount/extensions/orders/[id]/before'
 import { injectGlobalSections } from 'src/server/cms/global'
 
 const COMPONENTS: Record<string, ComponentType<any>> = {
@@ -35,9 +35,9 @@ export default function OrderDetailsPage({
       <NextSeo noindex nofollow />
 
       <MyAccountLayout>
-        <BeforeSession />
+        <BeforeSection />
         <MyAccountOrderDetails orderId={id} />
-        <AfterSession />
+        <AfterSection />
       </MyAccountLayout>
     </RenderSections>
   )
